NumericalSolutionOfPartialDifferentialEquations

Documentation for NumericalSolutionOfPartialDifferentialEquations.

Laplace Equation

NumericalSolutionOfPartialDifferentialEquations

Index

Index

Base.:*

NumericalSolutionOfPartialDifferentialEquations.add_boundary_condition

NumericalSolutionOfPartialDifferentialEquations.construct_boundary

NumericalSolutionOfPartialDifferentialEquations.construct_element_stiffness_matrix_with_abs_det_A_e_2

NumericalSolutionOfPartialDifferentialEquations.construct_grid

NumericalSolutionOfPartialDifferentialEquations.construct_laplacian

NumericalSolutionOfPartialDifferentialEquations.homework1

NumericalSolutionOfPartialDifferentialEquations.plot_1d_heat_equ_solution

NumericalSolutionOfPartialDifferentialEquations.solve_heat_equation

NumericalSolutionOfPartialDifferentialEquations.solve_heat_equation_explicit

NumericalSolutionOfPartialDifferentialEquations.solve_poissions_equation

NumericalSolutionOfPartialDifferentialEquations.test_solve_heat_equation

NumericalSolutionOfPartialDifferentialEquations.test_solve_poissions_equation_known

NumericalSolutionOfPartialDifferentialEquations.test_solve_poissions_equation_unknown

Base.:* — Method

Default operation overloads for I (identity matrix) are too slow, so I write these overloads.

NumericalSolutionOfPartialDifferentialEquations.add_boundary_condition — Method

add_boundary_condition(
    f::Vector{T1},
    boundary::Vector{Vector{T2}},
) where {T1<:Number,T2<:Number}

Add Dirichlet boundary conditions to the right of difference equation.

Left:   boundary[1]
Bottom: boundary[2]
Right:  boundary[3]
Top:    boundary[4]

They are column vectors and the orientation is counterclockwise.

NumericalSolutionOfPartialDifferentialEquations.construct_boundary — Method

construct_boundary(
    bound_func::Vector{T},
    len::Int,
    wid::Int = len,
) where {T<:Function}

Use bound_func to construct boundary conditions.

NumericalSolutionOfPartialDifferentialEquations.construct_element_stiffness_matrix_with_abs_det_A_e_2 — Method

\[A^e = \begin{pmatrix} x_1^1 & x_1^2 & x_1^3 \\ x_2^1 & x_2^2 & x_2^3 \end{pmatrix}\]

NumericalSolutionOfPartialDifferentialEquations.construct_grid — Method

construct_grid(f::T, len::Int, wid::Int = len) where {T<:Function}

Use f to construct grid.

NumericalSolutionOfPartialDifferentialEquations.construct_laplacian — Method

construct_laplacian(; size::Int = 32, dim::Int = 2)

Construct Laplacian by this method:

\[\boldsymbol L_{\mathrm{dim}} = \sum_{i = 1}^{m} \left( \bigotimes_{j = 1}^{i - 1} \boldsymbol I_n \right) \otimes \boldsymbol L_1 \otimes \left(\bigotimes_{j = i + 1}^{m} \boldsymbol I_n\right)\]

NumericalSolutionOfPartialDifferentialEquations.homework1 — Method

Need ~500 GiB Memory!!!

NumericalSolutionOfPartialDifferentialEquations.plot_1d_heat_equ_solution — Method

Also used by convection diffusion equation

NumericalSolutionOfPartialDifferentialEquations.solve_heat_equation — Method

solve_heat_equation(
    thermal_diffusivity::T1,
    initial::Vector{T2};
    solver::Function = \,
    total_time::T3 = 1.0,
    time_step_num::Int = 1024,
    theta::T4 = 0.5,
    kw...,
) where {T1<:Number,T2<:Number,T3<:Number,T4<:Number}

Solve Heat equation with Dirichlet boundary conditions (zero boundary and given initial).

NumericalSolutionOfPartialDifferentialEquations.solve_heat_equation_explicit — Method

Three Scheme

NumericalSolutionOfPartialDifferentialEquations.solve_poissions_equation — Method

solve_poissions_equation(
    f::Vector{T1},
    boundary::Vector{Vector{T2}},
) where {T1<:Number,T2<:Number}

Solve Poisson's equation with Dirichlet boundary conditions.

NumericalSolutionOfPartialDifferentialEquations.test_solve_heat_equation — Method

\[(x, t) \in (0, 1)^2\]

NumericalSolutionOfPartialDifferentialEquations.test_solve_poissions_equation_known — Method

# 已知精确解
# 共轭梯度?LU分解?

NumericalSolutionOfPartialDifferentialEquations.test_solve_poissions_equation_unknown — Method

# 无精确解 log-log plot